test(react-utils): ensure passing tests
