+++ packages/menu/src/use-menu.ts
@@ -25,10 +25,9 @@ import {
   getValidChildren,
   isArray,
   isString,
-  normalizeEventKey,
   removeItem,
 } from "@chakra-ui/utils"
-import { mergeRefs } from "react-utils"
+import { mergeRefs, normalizeEventKey } from "@chakra-ui/react-utils"
 import * as React from "react"
 
 const [MenuProvider, useMenuContext] = createContext<UseMenuReturn>({
