+++ packages/react-utils/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./use-react-utils"
 export * from "./react-utils"
+export * from "./dom"
