+++ packages/react-utils/tests/dom.test.ts
@@ -1,0 +1,6 @@
+import { normalizeEventKey } from "../src"
+
+test("should normalize keyboard events", () => {
+  const keyboardEvent: any = { key: "Left", keyCode: 38 }
+  expect(normalizeEventKey(keyboardEvent)).toBe("ArrowLeft")
+})
