+++ packages/react-utils/tests/react-utils.test.ts
@@ -1,3 +1,5 @@
-test("adds 1 + 2 to equal 3", () => {
-  expect(1 + 2).toBe(3)
+import { isInputEvent } from "../src"
+
+test("is input event", () => {
+  expect(isInputEvent({ target: {} })).toBeTruthy()
 })
