+++ packages/slider/src/use-slider.ts
@@ -18,14 +18,16 @@ import {
   focus,
   getBox,
   getOwnerDocument,
-  isRightClick,
-  normalizeEventKey,
   percentToValue,
   PropGetter,
   roundValueToStep,
   valueToPercent,
 } from "@chakra-ui/utils"
-import { mergeRefs } from "@chakra-ui/react-utils"
+import {
+  mergeRefs,
+  normalizeEventKey,
+  isRightClick,
+} from "@chakra-ui/react-utils"
 import { CSSProperties, useCallback, useMemo, useRef, useState } from "react"
 
 export interface UseSliderProps {
