+++ packages/utils/src/index.ts
@@ -7,7 +7,6 @@ export * from "./number"
 export * from "./dom-query"
 export * from "./tabbable"
 export * from "./assertion"
-export * from "./react-helpers"
 export * from "css-box-model"
 export * from "./responsive"
 export * from "./focus"
